Minor refactor: genome_pars_to_paths now in utils, and SOPRANOError in objects
